requirements.txt
dropbox~=12.0.2
psutil~=6.1.0
python-dotenv~=1.0.1
win10toast~=0.9